requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.11
reportlab==4.0.8
python-decouple==3.8